memory leaks fixed

+++ app/src/main/java/com/blongdev/sift/AccountInfo.java
@@ -4,10 +4,10 @@ package com.blongdev.sift;
  * Created by Brian on 2/24/2016.
  */
 public class AccountInfo {
-    public int mId;
+    public long mId;
     public String mRefreshKey;
     public String mUsername;
-    public int mUserId;
+    public long mUserId;
     public int mDateCreated;
     public String mPassword;
 }
