+++ app/src/main/java/com/blongdev/sift/CommentInfo.java
@@ -7,7 +7,7 @@ import net.dean.jraw.models.Comment;
  */
 public class CommentInfo extends ContributionInfo {
 
-    public int mParentId;
+    public long mParentId;
     public String mPostServerId;
     public Comment mJrawComment;
