+++ app/src/main/java/com/blongdev/sift/FavoritesInfo.java
@@ -4,6 +4,6 @@ package com.blongdev.sift;
  * Created by Brian on 2/24/2016.
  */
 public class FavoritesInfo {
-    public int mAccountId;
-    public int mPostId;
+    public long mAccountId;
+    public long mPostId;
 }
