+++ app/src/main/java/com/blongdev/sift/FriendInfo.java
@@ -4,8 +4,8 @@ package com.blongdev.sift;
  * Created by Brian on 2/24/2016.
  */
 public class FriendInfo {
-    public int mAccountId;
-    public int mFriendId;
+    public long mAccountId;
+    public long mFriendId;
 
     public String mFriendUsername;
     public int mFriendUserType;
