+++ app/src/main/java/com/blongdev/sift/MessageInfo.java
@@ -4,11 +4,11 @@ package com.blongdev.sift;
  * Created by Brian on 2/24/2016.
  */
 public class MessageInfo {
-    public int mAccountId;
+    public long mAccountId;
     public String mTo;
-    public int mToUserId;
+    public long mToUserId;
     public String mFrom;
-    public int mFromUserId;
+    public long mFromUserId;
     public String mTitle;
     public String mBody;
     public int mDate;
