+++ app/src/main/java/com/blongdev/sift/SubscriptionInfo.java
@@ -4,7 +4,7 @@ package com.blongdev.sift;
  * Created by Brian on 2/24/2016.
  */
 public class SubscriptionInfo {
-    public int mAccountId;
+    public long mAccountId;
     public long mSubredditId;
     public String mServerId;
