+++ app/src/main/java/com/blongdev/sift/VoteInfo.java
@@ -4,8 +4,8 @@ package com.blongdev.sift;
  * Created by Brian on 2/24/2016.
  */
 public class VoteInfo {
-    public int mAccountId;
-    public int mPostId;
-    public int mCommentId;
+    public long mAccountId;
+    public long mPostId;
+    public long mCommentId;
     public int mVote;
 }
